feat: rename is.js to device.js
