+++ frontend_nuxt/utils/device.js
